pyproject.toml
[tool.poetry]
name = "My Cutie Kitten Bot"
version = "0.1.0"
description = ""
authors = ["Gio Borgovitz"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"
pymorphy3 = "^1.2.0"
attrs = "^23.1.0"
betterconf = "^2.6.1"
typing-extensions = "^4.7.1"
sqlalchemy = "^2.0.19"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
requirements-parser = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
